Move class to better namespace and minor cleanup
