+++ src/Npgsql/SqlQueryParser.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Text;
+using Npgsql.Util;
 
 namespace Npgsql
 {
