+++ src/Npgsql/Util/StringBuilderExtensions.cs
@@ -1,8 +1,8 @@
-#if (NETSTANDARD2_0 || NET452)
+#if NETSTANDARD2_0 || NET452
 using System;
 using System.Text;
 
-namespace Npgsql
+namespace Npgsql.Util
 {
     /// <summary>
     /// A set of extension methods to <see cref="StringBuilder"/> to allow runtime compatibility.
